ysplit/requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
click==7.1.1
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.4.1
PyYAML==5.3.1
regex==2020.2.20
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
wcwidth==0.1.9
